+++ app/client/src/widgets/ButtonWidget/widget/index.tsx
@@ -31,6 +31,149 @@ class ButtonWidget extends BaseWidget<ButtonWidgetProps, ButtonWidgetState> {
     };
   }
 
+  static getPropertyPaneContentConfig() {
+    return [
+      {
+        sectionName: "Basic",
+        children: [
+          {
+            propertyName: "text",
+            label: "Label",
+            helpText: "Sets the label of the button",
+            controlType: "INPUT_TEXT",
+            placeholderText: "Submit",
+            isBindProperty: true,
+            isTriggerProperty: false,
+            validation: { type: ValidationTypes.TEXT },
+          },
+          {
+            helpText: "Triggers an action when the button is clicked",
+            propertyName: "onClick",
+            label: "onClick",
+            controlType: "ACTION_SELECTOR",
+            isJSConvertible: true,
+            isBindProperty: true,
+            isTriggerProperty: true,
+          },
+        ],
+      },
+      {
+        sectionName: "General",
+        children: [
+          {
+            helpText: "Show helper text with button on hover",
+            propertyName: "tooltip",
+            label: "Tooltip",
+            controlType: "INPUT_TEXT",
+            placeholderText: "Submits Form",
+            isBindProperty: true,
+            isTriggerProperty: false,
+            validation: { type: ValidationTypes.TEXT },
+          },
+          {
+            propertyName: "isVisible",
+            label: "Visible",
+            helpText: "Controls the visibility of the widget",
+            controlType: "SWITCH",
+            isJSConvertible: true,
+            isBindProperty: true,
+            isTriggerProperty: false,
+            validation: { type: ValidationTypes.BOOLEAN },
+          },
+          {
+            propertyName: "isDisabled",
+            label: "Disabled",
+            controlType: "SWITCH",
+            helpText: "Disables clicks to this widget",
+            isJSConvertible: true,
+            isBindProperty: true,
+            isTriggerProperty: false,
+            validation: { type: ValidationTypes.BOOLEAN },
+          },
+          {
+            propertyName: "animateLoading",
+            label: "Animate Loading",
+            controlType: "SWITCH",
+            helpText: "Controls the loading of the widget",
+            defaultValue: true,
+            isJSConvertible: true,
+            isBindProperty: true,
+            isTriggerProperty: false,
+            validation: { type: ValidationTypes.BOOLEAN },
+          },
+        ],
+      },
+      {
+        sectionName: "Validation",
+        children: [
+          {
+            propertyName: "googleRecaptchaKey",
+            label: "Google reCAPTCHA Key",
+            helpText: "Sets Google reCAPTCHA site key for the button",
+            controlType: "INPUT_TEXT",
+            placeholderText: "reCAPTCHA Key",
+            isBindProperty: true,
+            isTriggerProperty: false,
+            validation: { type: ValidationTypes.TEXT },
+          },
+          {
+            propertyName: "recaptchaType",
+            label: "Google reCAPTCHA Version",
+            controlType: "DROP_DOWN",
+            helpText: "Select reCAPTCHA version",
+            options: [
+              {
+                label: "reCAPTCHA v3",
+                value: RecaptchaTypes.V3,
+              },
+              {
+                label: "reCAPTCHA v2",
+                value: RecaptchaTypes.V2,
+              },
+            ],
+            isBindProperty: true,
+            isTriggerProperty: false,
+            validation: {
+              type: ValidationTypes.TEXT,
+              params: {
+                allowedValues: [RecaptchaTypes.V3, RecaptchaTypes.V2],
+                default: RecaptchaTypes.V3,
+              },
+            },
+          },
+        ],
+      },
+      // TODO: refactor widgetParentProps implementation when we address #10659
+      {
+        sectionName: "Form Settings",
+        children: [
+          {
+            helpText:
+              "Disabled if the form is invalid, if this widget exists directly within a Form widget.",
+            propertyName: "disabledWhenInvalid",
+            label: "Disabled Invalid Forms",
+            controlType: "SWITCH",
+            isJSConvertible: true,
+            isBindProperty: true,
+            isTriggerProperty: false,
+            validation: { type: ValidationTypes.BOOLEAN },
+          },
+          {
+            helpText:
+              "Resets the fields of the form, on click, if this widget exists directly within a Form widget.",
+            propertyName: "resetFormOnClick",
+            label: "Reset Form on Success",
+            controlType: "SWITCH",
+            isJSConvertible: true,
+            isBindProperty: true,
+            isTriggerProperty: false,
+            validation: { type: ValidationTypes.BOOLEAN },
+          },
+        ],
+      },
+    ];
+  }
+
   static getPropertyPaneStyleConfig() {
     return [
       {
@@ -203,302 +346,6 @@ class ButtonWidget extends BaseWidget<ButtonWidgetProps, ButtonWidgetState> {
             isBindProperty: true,
             isTriggerProperty: false,
             validation: { type: ValidationTypes.TEXT },
-          },
-        ],
-      },
-    ];
-  }
-
-  static getPropertyPaneConfig() {
-    return [
-      {
-        sectionName: "General",
-        children: [
-          {
-            propertyName: "text",
-            label: "Label",
-            helpText: "Sets the label of the button",
-            controlType: "INPUT_TEXT",
-            placeholderText: "Submit",
-            isBindProperty: true,
-            isTriggerProperty: false,
-            validation: { type: ValidationTypes.TEXT },
-          },
-          {
-            helpText: "Show helper text with button on hover",
-            propertyName: "tooltip",
-            label: "Tooltip",
-            controlType: "INPUT_TEXT",
-            placeholderText: "Submits Form",
-            isBindProperty: true,
-            isTriggerProperty: false,
-            validation: { type: ValidationTypes.TEXT },
-          },
-          {
-            propertyName: "googleRecaptchaKey",
-            label: "Google reCAPTCHA Key",
-            helpText: "Sets Google reCAPTCHA site key for the button",
-            controlType: "INPUT_TEXT",
-            placeholderText: "reCAPTCHA Key",
-            isBindProperty: true,
-            isTriggerProperty: false,
-            validation: { type: ValidationTypes.TEXT },
-          },
-          {
-            propertyName: "recaptchaType",
-            label: "Google reCAPTCHA Version",
-            controlType: "DROP_DOWN",
-            helpText: "Select reCAPTCHA version",
-            options: [
-              {
-                label: "reCAPTCHA v3",
-                value: RecaptchaTypes.V3,
-              },
-              {
-                label: "reCAPTCHA v2",
-                value: RecaptchaTypes.V2,
-              },
-            ],
-            isBindProperty: true,
-            isTriggerProperty: false,
-            validation: {
-              type: ValidationTypes.TEXT,
-              params: {
-                allowedValues: [RecaptchaTypes.V3, RecaptchaTypes.V2],
-                default: RecaptchaTypes.V3,
-              },
-            },
-          },
-          {
-            propertyName: "isVisible",
-            label: "Visible",
-            helpText: "Controls the visibility of the widget",
-            controlType: "SWITCH",
-            isJSConvertible: true,
-            isBindProperty: true,
-            isTriggerProperty: false,
-            validation: { type: ValidationTypes.BOOLEAN },
-          },
-          {
-            propertyName: "isDisabled",
-            label: "Disabled",
-            controlType: "SWITCH",
-            helpText: "Disables clicks to this widget",
-            isJSConvertible: true,
-            isBindProperty: true,
-            isTriggerProperty: false,
-            validation: { type: ValidationTypes.BOOLEAN },
-          },
-          {
-            propertyName: "animateLoading",
-            label: "Animate Loading",
-            controlType: "SWITCH",
-            helpText: "Controls the loading of the widget",
-            defaultValue: true,
-            isJSConvertible: true,
-            isBindProperty: true,
-            isTriggerProperty: false,
-            validation: { type: ValidationTypes.BOOLEAN },
-          },
-        ],
-      },
-      // TODO: refactor widgetParentProps implementation when we address #10659
-      {
-        sectionName: "Form options",
-        children: [
-          {
-            helpText:
-              "Disabled if the form is invalid, if this widget exists directly within a Form widget.",
-            propertyName: "disabledWhenInvalid",
-            label: "Disabled Invalid Forms",
-            controlType: "SWITCH",
-            isJSConvertible: true,
-            isBindProperty: true,
-            isTriggerProperty: false,
-            validation: { type: ValidationTypes.BOOLEAN },
-          },
-          {
-            helpText:
-              "Resets the fields of the form, on click, if this widget exists directly within a Form widget.",
-            propertyName: "resetFormOnClick",
-            label: "Reset Form on Success",
-            controlType: "SWITCH",
-            isJSConvertible: true,
-            isBindProperty: true,
-            isTriggerProperty: false,
-            validation: { type: ValidationTypes.BOOLEAN },
-          },
-        ],
-      },
-      {
-        sectionName: "Events",
-        children: [
-          {
-            helpText: "Triggers an action when the button is clicked",
-            propertyName: "onClick",
-            label: "onClick",
-            controlType: "ACTION_SELECTOR",
-            isJSConvertible: true,
-            isBindProperty: true,
-            isTriggerProperty: true,
-          },
-        ],
-      },
-      {
-        sectionName: "Styles",
-        children: [
-          {
-            propertyName: "buttonColor",
-            helpText: "Changes the color of the button",
-            label: "Button Color",
-            controlType: "COLOR_PICKER",
-            isJSConvertible: true,
-            isBindProperty: true,
-            isTriggerProperty: false,
-            validation: { type: ValidationTypes.TEXT },
-          },
-          {
-            propertyName: "buttonVariant",
-            label: "Button Variant",
-            controlType: "DROP_DOWN",
-            helpText: "Sets the variant of the icon button",
-            options: [
-              {
-                label: "Primary",
-                value: ButtonVariantTypes.PRIMARY,
-              },
-              {
-                label: "Secondary",
-                value: ButtonVariantTypes.SECONDARY,
-              },
-              {
-                label: "Tertiary",
-                value: ButtonVariantTypes.TERTIARY,
-              },
-            ],
-            isJSConvertible: true,
-            isBindProperty: true,
-            isTriggerProperty: false,
-            validation: {
-              type: ValidationTypes.TEXT,
-              params: {
-                allowedValues: [
-                  ButtonVariantTypes.PRIMARY,
-                  ButtonVariantTypes.SECONDARY,
-                  ButtonVariantTypes.TERTIARY,
-                ],
-                default: ButtonVariantTypes.PRIMARY,
-              },
-            },
-          },
-          {
-            propertyName: "borderRadius",
-            label: "Border Radius",
-            helpText:
-              "Rounds the corners of the icon button's outer border edge",
-            controlType: "BORDER_RADIUS_OPTIONS",
-            isBindProperty: true,
-            isJSConvertible: true,
-            isTriggerProperty: false,
-            validation: {
-              type: ValidationTypes.TEXT,
-            },
-          },
-          {
-            propertyName: "boxShadow",
-            label: "Box Shadow",
-            helpText:
-              "Enables you to cast a drop shadow from the frame of the widget",
-            controlType: "BOX_SHADOW_OPTIONS",
-            isJSConvertible: true,
-            isBindProperty: true,
-            isTriggerProperty: false,
-            validation: { type: ValidationTypes.TEXT },
-          },
-          {
-            propertyName: "iconName",
-            label: "Icon",
-            helpText: "Sets the icon to be used for the button",
-            controlType: "ICON_SELECT",
-            isBindProperty: false,
-            isTriggerProperty: false,
-            updateHook: (
-              props: ButtonWidgetProps,
-              propertyPath: string,
-              propertyValue: string,
-            ) => {
-              const propertiesToUpdate = [{ propertyPath, propertyValue }];
-              if (!props.iconAlign) {
-                propertiesToUpdate.push({
-                  propertyPath: "iconAlign",
-                  propertyValue: Alignment.LEFT,
-                });
-              }
-              return propertiesToUpdate;
-            },
-            dependencies: ["iconAlign"],
-            validation: {
-              type: ValidationTypes.TEXT,
-            },
-          },
-          {
-            propertyName: "placement",
-            label: "Placement",
-            controlType: "DROP_DOWN",
-            helpText: "Sets the space between items",
-            options: [
-              {
-                label: "Start",
-                value: ButtonPlacementTypes.START,
-              },
-              {
-                label: "Between",
-                value: ButtonPlacementTypes.BETWEEN,
-              },
-              {
-                label: "Center",
-                value: ButtonPlacementTypes.CENTER,
-              },
-            ],
-            defaultValue: ButtonPlacementTypes.CENTER,
-            isJSConvertible: true,
-            isBindProperty: true,
-            isTriggerProperty: false,
-            validation: {
-              type: ValidationTypes.TEXT,
-              params: {
-                allowedValues: [
-                  ButtonPlacementTypes.START,
-                  ButtonPlacementTypes.BETWEEN,
-                  ButtonPlacementTypes.CENTER,
-                ],
-                default: ButtonPlacementTypes.CENTER,
-              },
-            },
-          },
-          {
-            propertyName: "iconAlign",
-            label: "Icon Alignment",
-            helpText: "Sets the icon alignment of the button",
-            controlType: "ICON_TABS",
-            options: [
-              {
-                icon: "VERTICAL_LEFT",
-                value: "left",
-              },
-              {
-                icon: "VERTICAL_RIGHT",
-                value: "right",
-              },
-            ],
-            isBindProperty: false,
-            isTriggerProperty: false,
-            validation: {
-              type: ValidationTypes.TEXT,
-              params: {
-                allowedValues: ["center", "left", "right"],
-              },
-            },
           },
         ],
       },
